improve string32 flag check and use tests/main.cpp as combined test executable
