+++ UTD/src/data-structures/String.cpp
@@ -1,4 +1,4 @@
-#include "String.h"
+#include "string.h"
 
 #include <cstring>	// should this be removed?
